# lib/api-zod/src/generated/types/callRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PeriOp Voice API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CallRecordCallType } from './callRecordCallType';
import type { CallRecordOutcome } from './callRecordOutcome';
import type { Patient } from './patient';

export interface CallRecord {
  id: number;
  scheduledCallId: number;
  patientId: number;
  callType: CallRecordCallType;
  outcome?: CallRecordOutcome;
  /** @nullable */
  durationSeconds?: number | null;
  startedAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  /** @nullable */
  aiSummary?: string | null;
  hasFlags?: boolean;
  createdAt: Date;
  patient?: Patient;
}